Sepparated Session class and derivatives to their own files
